Implement order management functionality; add Order, OrderRequest, OrderService, OrderRepository, OrderController, and AdminOrderController classes

+++ Online-Food-ordering/src/main/java/com/sandarun/Online/Food/ordering/model/Order.java
@@ -41,5 +41,5 @@ public class Order {
 //    private Payment payment;
 
     private  int totalItem;
-    private  int totalPrice;
+    private  Long totalPrice;
 }
